Move theme environment files to their own folder
