+++ packages/theme/src/cli/utilities/theme-environment/asset-file-syncer.ts
@@ -1,4 +1,4 @@
-import {uploadTheme} from './theme-uploader.js'
+import {uploadTheme} from '../theme-uploader.js'
 import {outputDebug} from '@shopify/cli-kit/node/output'
 import {AdminSession} from '@shopify/cli-kit/node/session'
 import {Checksum, Theme, ThemeAsset, ThemeFileSystem} from '@shopify/cli-kit/node/themes/types'
